+++ crates/swc_ecma_minifier/src/compress/optimize/sequences.rs
@@ -1,5 +1,6 @@
 use std::mem::take;
 
+#[allow(unused_imports)]
 use retain_mut::RetainMut;
 use swc_atoms::js_word;
 use swc_common::{util::take::Take, Spanned, DUMMY_SP};