+++ crates/swc_ecma_parser/src/parser/expr.rs
@@ -12,7 +12,7 @@ mod ops;
 mod tests;
 mod verifier;
 
-impl<'a, I: Tokens> Parser<I> {
+impl<I: Tokens> Parser<I> {
     pub fn parse_expr(&mut self) -> PResult<Box<Expr>> {
         trace_cur!(self, parse_expr);
 
@@ -1800,7 +1800,7 @@ pub(in crate::parser) enum PatOrExprOrSpread {
 
 /// simple leaf methods.
 
-impl<'a, I: Tokens> Parser<I> {
+impl<I: Tokens> Parser<I> {
     fn parse_yield_expr(&mut self) -> PResult<Box<Expr>> {
         let start = cur_pos!(self);